Prophet_/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cmdstanpy==1.3.0
contourpy==1.3.3
cycler==0.12.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.121.3
fonttools==4.60.1
h11==0.16.0
holidays==0.85
idna==3.11
importlib_resources==6.5.2
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.0
numpy==2.3.5
nvidia-nccl-cu12==2.28.9
packaging==25.0
pandas==2.2.3
pillow==12.0.0
prophet==1.2.1
pydantic==2.12.4
pydantic_core==2.41.5
pyparsing==3.2.5
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
sniffio==1.3.1
stanio==0.5.1
starlette==0.50.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
xgboost==2.1.4
